Finished transfer methods, and button click events
Added checkingToSavings() which transfers the given amount of money from the checking account to the savings account. savingsToChecking() accomplishes the same thing, vice versa. A new JButtton loginExitButton was added to properly exit the program. Button click events were cleaned up and added as well. checkToSaveButtonClicked() starts the process of checking for appropriate input, and calling the public method in bankhandlder, eventually updating the labels on the panel to reflect the new current balances. Same goes for saveToCheckButtonClicked(), but reversed